Feature ETP-3674: Extract WebhookBaseTest to remove code duplication
